Moved unit tests to __tests__/
