+++ __tests__/pages/home.test.js
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 import { render, screen } from "@testing-library/react";
-import Home from "./home";
+import Home from "../../pages/home";
 
 describe("Home", () => {
   it("renders without crashing", () => {
