+++ __tests__/pages/signup.test.js
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 import { render, screen } from "@testing-library/react";
-import Signup from "./signup";
+import Signup from "../../pages/signup";
 
 describe("Signup", () => {
   it("renders without crashing", () => {
