+++ __tests__/pages/thankyou.test.js
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 import { render, screen } from "@testing-library/react";
-import ThankYou from "./thankyou";
+import ThankYou from "../../pages/thankyou";
 
 describe("ThankYou", () => {
   it("renders without crashing", () => {
